bugfix for torch 1.6.0
torch 1.6.0 doesn't like using / for division between int tensors, we
should either use // or cast them to floats. in this commit, we cast
them to floats

+++ plenoptic/simulate/models/texture_statistics.py
@@ -128,7 +128,7 @@ class Texture_Statistics(nn.Module):
         im = mpyr[0].squeeze()
 
         # Find the auto-correlation of the low-pass residual
-        Sch = torch.min(torch.tensor(ch.shape[-2:]))
+        Sch = torch.min(torch.tensor(ch.shape[-2:])).to(float)
         la = int(np.floor([(self.Na-1)/2]))
         le = int(np.min((Sch/2-1,la)))
         acr[la-le:la+le+1, la-le:la+le+1, self.n_scales], vari = self.compute_autocorr(im)
@@ -141,7 +141,7 @@ class Texture_Statistics(nn.Module):
             for nor in range(0, self.n_orientations):
                 nband = n_scales*self.n_orientations + nor + 1
                 ch = apyr0[nband]
-                Sch = np.min((ch.shape[-1], ch.shape[-2]))
+                Sch = np.min((ch.shape[-1], ch.shape[-2])).to(float)
                 le = int(np.min((Sch/2-1, la)))
                 # Find the auto-correlation of the magnitude band
                 ace[la-le:la+le+1, la-le:la+le+1, n_scales, nor], vari = self.compute_autocorr(ch)
@@ -314,7 +314,7 @@ class Texture_Statistics(nn.Module):
 
     def compute_autocorr(self,ch):
 
-        Sch = torch.min(torch.tensor(ch.shape[-2:]))
+        Sch = torch.min(torch.tensor(ch.shape[-2:])).to(float)
 
         la = int(np.floor([(self.Na-1)/2]))
         le = int(np.min((Sch/2-1,la)))
